# Backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FirstRegistrars BuildWise API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [internal, vendor]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- TASKS ----
  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: sprintId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: assigneeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /projects/{projectId}/sprints:
    get:
      operationId: listSprints
      tags: [sprints]
      summary: List sprints for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sprints list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sprint"
    post:
      operationId: createSprint
      tags: [sprints]
      summary: Create a sprint
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSprintBody"
      responses:
        "201":
          description: Created sprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
...
  /ai/analyze-project:
    post:
      operationId: analyzeProject
      tags: [ai]
      summary: Analyze a project with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeProjectBody"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysisResult"
  /ai/business-advice:
    post:
      operationId: getBusinessAdvice
      tags: [ai]
      summary: Get AI business advice for a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessAdviceBody"
      responses:
        "200":
          description: Business advice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysisResult"
  /ai/version-advice:
    post:
      operationId: getVersionAdvice
      tags: [ai]
      summary: Get AI advice on when to release a new version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VersionAdviceBody"
      responses:
        "200":
          description: Version advice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysisResult"
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [internal, vendor]
        status:
          type: string
          enum: [planning, in_progress, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        country:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        budget:
          type: number
          nullable: true
        completionRate:
          type: number
        ownerId:
          type: integer
          nullable: true
        vendorId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - status
        - priority
        - completionRate
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [internal, vendor]
        status:
          type: string
          enum: [planning, in_progress, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        country:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        budget:
          type: number
          nullable: true
        ownerId:
          type: integer
          nullable: true
        vendorId:
          type: integer
          nullable: true
      required:
        - name
        - type
        - status
        - priority
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [internal, vendor]
        status:
          type: string
          enum: [planning, in_progress, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        country:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        budget:
          type: number
          nullable: true
        completionRate:
          type: number
        ownerId:
          type: integer
          nullable: true
        vendorId:
          type: integer
          nullable: true

    # TASKS
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        sprintId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          enum: [story, task, bug, epic, subtask]
        assigneeId:
          type: integer
          nullable: true
        reporterId:
          type: integer
          nullable: true
        storyPoints:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        label:
          type: string
          nullable: true
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - status
        - priority
        - type
        - position
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        sprintId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, critical]
        type:
          type: string
          enum: [story, task, bug, epic, subtask]
        assigneeId:
          type: integer
          nullable: true
        reporterId:
          type: integer
          nullable: true
        storyPoints:
          type: integer
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        label:
          type: string
          nullable: true
      required:
        - title
        - status
        - priority
        - type
...
    Sprint:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        goal:
          type: string
          nullable: true
        status:
          type: string
          enum: [planned, active, completed]
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - status
        - createdAt
    CreateSprintBody:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
          nullable: true
        status:
          type: string
          enum: [planned, active, completed]
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
      required:
        - name
        - status
...
    AnalyzeProjectBody:
      type: object
      properties:
        projectId:
          type: integer
        includeFinancial:
          type: boolean
      required:
        - projectId
    BusinessAdviceBody:
      type: object
      properties:
        projectId:
          type: integer
        country:
          type: string
        question:
          type: string
          nullable: true
      required:
        - projectId
        - country
    VersionAdviceBody:
      type: object
      properties:
        projectId:
          type: integer
      required:
        - projectId
    AiAnalysisResult:
      type: object
      properties:
        summary:
          type: string
        completionRateAnalysis:
          type: string
          nullable: true
        profitabilityScore:
          type: number
          nullable: true
        recommendation:
          type: string
          enum: [continue, pause, stop, expand, review]
          nullable: true
        insights:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        versionAdvice:
          type: string
          nullable: true
        countryAnalysis:
          type: string
          nullable: true
      required:
        - summary
        - insights
        - suggestions
        - risks