# Backend/lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FirstRegistrars BuildWise API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectBodyPriority } from "./createProjectBodyPriority";
import type { CreateProjectBodyStatus } from "./createProjectBodyStatus";
import type { CreateProjectBodyType } from "./createProjectBodyType";

export interface CreateProjectBody {
  name: string;
  description?: string | null;
  type: CreateProjectBodyType;
  status: CreateProjectBodyStatus;
  priority: CreateProjectBodyPriority;
  country?: string | null;
  startDate?: Date | null;
  endDate?: Date | null;
  budget?: number | null;
  ownerId?: number | null;
  vendorId?: number | null;
}