# Backend/lib/db/src/schema/projects.ts
import { pgTable, serial, text, timestamp, integer, numeric, date } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { usersTable } from "./users";
import { vendorsTable } from "./vendors";

export const projectsTable = pgTable("projects", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  description: text("description"),
  type: text("type").notNull().default("internal"),
  status: text("status").notNull().default("planning"),
  priority: text("priority").notNull().default("medium"),
  country: text("country"),
  startDate: date("start_date"),
  endDate: date("end_date"),
  budget: numeric("budget", { precision: 15, scale: 2 }),
  completionRate: numeric("completion_rate", { precision: 5, scale: 2 }).notNull().default("0"),
  ownerId: integer("owner_id").references(() => usersTable.id),
  vendorId: integer("vendor_id").references(() => vendorsTable.id),
  createdAt: timestamp("created_at").notNull().defaultNow(),
  updatedAt: timestamp("updated_at").notNull().defaultNow(),
});

export const insertProjectSchema = createInsertSchema(projectsTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertProject = z.infer<typeof insertProjectSchema>;
export type Project = typeof projectsTable.$inferSelect;